feat(graph): extract shared source-signal eligibility module
Move extractFrontmatter, parseSourcesFrontmatter, and related helpers
from graph-analysis.js into scripts/lib/source-signal-eligibility.js.
Add evaluateSourceSignalEligibility() with shared page-type mapping
and SCAN_KINDS constant. graph-analysis.js now requires the shared
module instead of maintaining its own copy.
